+++ backend/gonzales/static/assets/HistoryPage-B1zsZkxT.js
@@ -1,4 +1,4 @@
-import{a as i,j as e}from"./query-CSQiZfg4.js";import{c as D,T as C,X as E,j as w,d as T,k as A,l as S,m as L,S as z}from"./index-CRbZuQtx.js";import{G as F}from"./GlassCard-D-K6o67A.js";import{D as U}from"./DateRangeFilter-BCNFgG6q.js";import{f as _}from"./format-Bm42Q-0N.js";import{A as P,a as R}from"./arrow-up-BDwBAm4u.js";import"./vendor-7Rr5RJ8N.js";import"./calendar-CMDwOKOH.js";/**
+import{a as i,j as e}from"./query-CSQiZfg4.js";import{c as D,T as C,X as E,j as w,d as T,k as A,l as S,m as L,S as z}from"./index-BCdNWXkU.js";import{G as F}from"./GlassCard-D-K6o67A.js";import{D as U}from"./DateRangeFilter-7-x5PQw2.js";import{f as _}from"./format-Bm42Q-0N.js";import{A as P,a as R}from"./arrow-up-CHRmhgxz.js";import"./vendor-7Rr5RJ8N.js";import"./calendar-DBFlsmmi.js";/**
  * @license lucide-react v0.563.0 - ISC
  *
  * This source code is licensed under the ISC license.