+++ backend/gonzales/static/assets/chevron-right-B5XhPT_o.js
@@ -1,4 +1,4 @@
-import{c as o}from"./index-CRbZuQtx.js";/**
+import{c as o}from"./index-BCdNWXkU.js";/**
  * @license lucide-react v0.563.0 - ISC
  *
  * This source code is licensed under the ISC license.